+++ crates/objc2/src/macros/declare_class.rs
@@ -93,11 +93,10 @@
 /// will be registered as a class method.
 ///
 /// On instance methods, you can freely choose between different types of
-/// receivers, e.g. `&self`, `this: *const Self`, `&mut self`, and so on. Note
-/// though that using `&mut self` requires the class' mutability to be
-/// [`IsAllowedMutable`].
-/// If you need mutation of your class' instance variables, consider using
-/// [`Cell`] or similar instead.
+/// receivers, e.g. `&self`, `self: *const Self`, `this: *const Self`, and so
+/// on. Note that using `&mut self` is not possible, if you need mutation of
+/// your class' instance variables, consider using [`Cell`] or similar
+/// instead.
 ///
 /// The desired selector can be specified using the `#[method(my:selector:)]`
 /// or `#[method_id(my:selector:)]` attribute, similar to the
@@ -127,7 +126,6 @@
 ///
 /// ["associated functions"]: https://doc.rust-lang.org/reference/items/associated-items.html#methods
 /// ["methods"]: https://doc.rust-lang.org/reference/items/associated-items.html#methods
-/// [`IsAllowedMutable`]: crate::mutability::IsAllowedMutable
 /// [`Cell`]: core::cell::Cell
 /// [open an issue]: https://github.com/madsmtm/objc2/issues/new
 /// [`msg_send!`]: crate::msg_send