+++ crates/objc2/CHANGELOG.md
@@ -53,6 +53,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `objc`).
 * **BREAKING**: Removed `AsMut` and `BorrowMut` implementations in
   `extern_class!` and `declare_class!`.
+* **BREAKING**: Removed `mutability` options `Immutable`, `Mutable`,
+  `ImmutableWithMutableSubclass` and `MutableWithImmutableSuperclass`, along
+  with the related `IsAllowedMutable` and `IsMutable` traits.
+* **BREAKING**: Removed `DerefMut` implementation for `Retained<T>` when the
+  `Retained` was mutable.
 
 ### Fixed
 * Remove an incorrect assertion when adding protocols to classes in an unexpected